+++ subscribers/market_depth_test.go
@@ -3,8 +3,10 @@ package subscribers_test
 import (
 	"context"
 	"testing"
+	"time"
 
 	"code.vegaprotocol.io/data-node/subscribers"
+	"code.vegaprotocol.io/protos/vega"
 	"code.vegaprotocol.io/vega/events"
 	"code.vegaprotocol.io/vega/types"
 	"code.vegaprotocol.io/vega/types/num"
@@ -31,24 +33,43 @@ func buildOrder(id string, side types.Side, orderType types.OrderType, price uin
 	return order
 }
 
+type OrderEventWithVegaTime struct {
+	events.Event
+	order    *vega.Order
+	vegaTime time.Time
+}
+
+func (oe OrderEventWithVegaTime) VegaTime() time.Time {
+	return oe.vegaTime
+}
+
+func (oe OrderEventWithVegaTime) Order() *vega.Order {
+	return oe.order
+}
+
+func newOrderEvent(ctx context.Context, o *types.Order) OrderEventWithVegaTime {
+	oe := events.NewOrderEvent(ctx, o)
+	return OrderEventWithVegaTime{oe, oe.Order(), time.Now()}
+}
+
 func TestBuyPriceLevels(t *testing.T) {
 	ctx := context.Background()
 	mdb := getTestMDB(t, ctx, true)
 
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 9, 9)
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 
 	order2 := buildOrder("Order2", types.SideBuy, types.OrderTypeLimit, 102, 7, 7)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	order3 := buildOrder("Order3", types.SideBuy, types.OrderTypeLimit, 101, 8, 8)
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 
 	order4 := buildOrder("Order4", types.SideBuy, types.OrderTypeLimit, 99, 10, 10)
-	event4 := events.NewOrderEvent(ctx, order4)
+	event4 := newOrderEvent(ctx, order4)
 	mdb.Push(event4)
 
 	assert.Equal(t, 4, mdb.GetBuyPriceLevels("M"))
@@ -73,19 +94,19 @@ func TestSellPriceLevels(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order1 := buildOrder("Order1", types.SideSell, types.OrderTypeLimit, 100, 9, 9)
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 
 	order2 := buildOrder("Order2", types.SideSell, types.OrderTypeLimit, 102, 7, 7)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	order3 := buildOrder("Order3", types.SideSell, types.OrderTypeLimit, 101, 8, 8)
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 
 	order4 := buildOrder("Order4", types.SideSell, types.OrderTypeLimit, 99, 10, 10)
-	event4 := events.NewOrderEvent(ctx, order4)
+	event4 := newOrderEvent(ctx, order4)
 	mdb.Push(event4)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -110,7 +131,7 @@ func TestAddOrderToEmptyBook(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	assert.Equal(t, 1, mdb.GetBuyPriceLevels("M"))
@@ -126,12 +147,12 @@ func TestCancelOrder(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	cancelorder := *order
 	cancelorder.Status = types.OrderStatusCancelled
-	event2 := events.NewOrderEvent(ctx, &cancelorder)
+	event2 := newOrderEvent(ctx, &cancelorder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -147,12 +168,12 @@ func TestStoppedOrder(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	cancelorder := *order
 	cancelorder.Status = types.OrderStatusStopped
-	event2 := events.NewOrderEvent(ctx, &cancelorder)
+	event2 := newOrderEvent(ctx, &cancelorder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -168,12 +189,12 @@ func TestExpiredOrder(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	cancelorder := *order
 	cancelorder.Status = types.OrderStatusExpired
-	event2 := events.NewOrderEvent(ctx, &cancelorder)
+	event2 := newOrderEvent(ctx, &cancelorder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -189,18 +210,18 @@ func TestAmendOrderPrice(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	order2 := buildOrder("Order2", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	// Amend the price to force a change in price level
 	amendorder := *order
 	amendorder.Price = num.NewUint(90)
 	amendorder.OriginalPrice = num.NewUint(90)
-	event3 := events.NewOrderEvent(ctx, &amendorder)
+	event3 := newOrderEvent(ctx, &amendorder)
 	mdb.Push(event3)
 
 	assert.Equal(t, 2, mdb.GetBuyPriceLevels("M"))
@@ -218,13 +239,13 @@ func TestAmendOrderVolumeUp(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	amendorder := *order
 	amendorder.Size = 20
 	amendorder.Remaining = 20
-	event2 := events.NewOrderEvent(ctx, &amendorder)
+	event2 := newOrderEvent(ctx, &amendorder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 1, mdb.GetBuyPriceLevels("M"))
@@ -240,13 +261,13 @@ func TestAmendOrderVolumeDown(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	amendorder := *order
 	amendorder.Size = 5
 	amendorder.Remaining = 5
-	event2 := events.NewOrderEvent(ctx, &amendorder)
+	event2 := newOrderEvent(ctx, &amendorder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 1, mdb.GetBuyPriceLevels("M"))
@@ -262,13 +283,13 @@ func TestAmendOrderVolumeDownToZero(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	amendorder := *order
 	amendorder.Size = 0
 	amendorder.Remaining = 0
-	event2 := events.NewOrderEvent(ctx, &amendorder)
+	event2 := newOrderEvent(ctx, &amendorder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -284,12 +305,12 @@ func TestPartialFill(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	pforder := *order
 	pforder.Remaining = 5
-	event2 := events.NewOrderEvent(ctx, &pforder)
+	event2 := newOrderEvent(ctx, &pforder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 1, mdb.GetBuyPriceLevels("M"))
@@ -307,7 +328,7 @@ func TestIOCPartialFill(t *testing.T) {
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 5)
 	order.Status = types.OrderStatusPartiallyFilled
 	order.TimeInForce = types.OrderTimeInForceIOC
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -323,13 +344,13 @@ func TestFullyFill(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	fforder := *order
 	fforder.Remaining = 0
 	fforder.Status = types.OrderStatusFilled
-	event2 := events.NewOrderEvent(ctx, &fforder)
+	event2 := newOrderEvent(ctx, &fforder)
 	mdb.Push(event2)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -346,7 +367,7 @@ func TestMarketOrder(t *testing.T) {
 
 	// market orders should not stay on the book
 	marketorder := buildOrder("Order1", types.SideBuy, types.OrderTypeMarket, 100, 10, 10)
-	event1 := events.NewOrderEvent(ctx, marketorder)
+	event1 := newOrderEvent(ctx, marketorder)
 	mdb.Push(event1)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -364,7 +385,7 @@ func TestFOKOrder(t *testing.T) {
 	// FOK orders do not stay on the book
 	fokorder := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
 	fokorder.TimeInForce = types.OrderTimeInForceFOK
-	event := events.NewOrderEvent(ctx, fokorder)
+	event := newOrderEvent(ctx, fokorder)
 	mdb.Push(event)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -382,7 +403,7 @@ func TestIOCOrder(t *testing.T) {
 	// IOC orders do not stay on the book
 	iocorder := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
 	iocorder.TimeInForce = types.OrderTimeInForceIOC
-	event := events.NewOrderEvent(ctx, iocorder)
+	event := newOrderEvent(ctx, iocorder)
 	mdb.Push(event)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -400,7 +421,7 @@ func TestRejectedOrder(t *testing.T) {
 	// Rejected orders should be ignored
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
 	order.Status = types.OrderStatusRejected
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -418,7 +439,7 @@ func TestInvalidOrder(t *testing.T) {
 	// Invalid orders should be ignored
 	order := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
 	order.Status = types.OrderStatusUnspecified
-	event := events.NewOrderEvent(ctx, order)
+	event := newOrderEvent(ctx, order)
 	mdb.Push(event)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -434,17 +455,17 @@ func TestPartialMatchOrders(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 	order2 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 8)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	order3 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 5)
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 	order4 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 1)
-	event4 := events.NewOrderEvent(ctx, order4)
+	event4 := newOrderEvent(ctx, order4)
 	mdb.Push(event4)
 
 	assert.Equal(t, 1, mdb.GetBuyPriceLevels("M"))
@@ -460,18 +481,18 @@ func TestFullyMatchOrders(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 	order2 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 8)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	order3 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 5)
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 	order4 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 100, 10, 0)
 	order4.Status = types.OrderStatusFilled
-	event4 := events.NewOrderEvent(ctx, order4)
+	event4 := newOrderEvent(ctx, order4)
 	mdb.Push(event4)
 
 	assert.Equal(t, 0, mdb.GetBuyPriceLevels("M"))
@@ -487,18 +508,18 @@ func TestRemovingPriceLevels(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 101, 10, 10)
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 	order2 := buildOrder("Order2", types.SideBuy, types.OrderTypeLimit, 100, 10, 10)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 	order3 := buildOrder("Order3", types.SideBuy, types.OrderTypeLimit, 102, 10, 10)
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 
 	order4 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 101, 10, 0)
 	order4.Status = types.OrderStatusFilled
-	event4 := events.NewOrderEvent(ctx, order4)
+	event4 := newOrderEvent(ctx, order4)
 	mdb.Push(event4)
 
 	assert.Equal(t, 2, mdb.GetBuyPriceLevels("M"))
@@ -514,7 +535,7 @@ func TestMarketDepthFields(t *testing.T) {
 	mdb := getTestMDB(t, ctx, true)
 
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 101, 10, 10)
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 
 	md, err := mdb.GetMarketDepth(ctx, "M", 0)
@@ -539,14 +560,14 @@ func TestParkingOrder(t *testing.T) {
 	// Create a valid and live pegged order
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 101, 10, 10)
 	order1.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 
 	// Park it
 	order2 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 0, 10, 10)
 	order2.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order2.Status = types.OrderStatusParked
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	md, err := mdb.GetMarketDepth(ctx, "M", 0)
@@ -561,7 +582,7 @@ func TestParkingOrder(t *testing.T) {
 	order3 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 101, 10, 10)
 	order3.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order3.Status = types.OrderStatusActive
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 
 	md2, err := mdb.GetMarketDepth(ctx, "M", 0)
@@ -581,7 +602,7 @@ func TestParkedOrder(t *testing.T) {
 	order1 := buildOrder("Order1", types.SideBuy, types.OrderTypeLimit, 101, 10, 10)
 	order1.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order1.Status = types.OrderStatusParked
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 
 	md, err := mdb.GetMarketDepth(ctx, "M", 0)
@@ -601,63 +622,63 @@ func TestParkedOrder2(t *testing.T) {
 	order1 := buildOrder("Pegged1", types.SideBuy, types.OrderTypeLimit, 0, 10, 10)
 	order1.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order1.Status = types.OrderStatusParked
-	event1 := events.NewOrderEvent(ctx, order1)
+	event1 := newOrderEvent(ctx, order1)
 	mdb.Push(event1)
 
 	// Create normal order
 	order2 := buildOrder("Normal1", types.SideBuy, types.OrderTypeLimit, 100, 1, 1)
-	event2 := events.NewOrderEvent(ctx, order2)
+	event2 := newOrderEvent(ctx, order2)
 	mdb.Push(event2)
 
 	// Unpark pegged order
 	order3 := buildOrder("Pegged1", types.SideBuy, types.OrderTypeLimit, 99, 10, 10)
 	order3.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order3.Status = types.OrderStatusActive
-	event3 := events.NewOrderEvent(ctx, order3)
+	event3 := newOrderEvent(ctx, order3)
 	mdb.Push(event3)
 
 	// Cancel normal order
 	order4 := buildOrder("Normal1", types.SideBuy, types.OrderTypeLimit, 100, 1, 1)
 	order4.Status = types.OrderStatusCancelled
-	event4 := events.NewOrderEvent(ctx, order4)
+	event4 := newOrderEvent(ctx, order4)
 	mdb.Push(event4)
 
 	// Park pegged order
 	order5 := buildOrder("Pegged1", types.SideBuy, types.OrderTypeLimit, 99, 10, 10)
 	order5.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order5.Status = types.OrderStatusParked
-	event5 := events.NewOrderEvent(ctx, order5)
+	event5 := newOrderEvent(ctx, order5)
 	mdb.Push(event5)
 
 	// Create normal order
 	order6 := buildOrder("Normal2", types.SideBuy, types.OrderTypeLimit, 100, 1, 1)
-	event6 := events.NewOrderEvent(ctx, order6)
+	event6 := newOrderEvent(ctx, order6)
 	mdb.Push(event6)
 
 	// Unpark pegged order
 	order7 := buildOrder("Pegged1", types.SideBuy, types.OrderTypeLimit, 99, 10, 10)
 	order7.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order7.Status = types.OrderStatusActive
-	event7 := events.NewOrderEvent(ctx, order7)
+	event7 := newOrderEvent(ctx, order7)
 	mdb.Push(event7)
 
 	// Fill normal order
 	order8 := buildOrder("Normal2", types.SideBuy, types.OrderTypeLimit, 100, 1, 0)
 	order8.Status = types.OrderStatusFilled
-	event8 := events.NewOrderEvent(ctx, order8)
+	event8 := newOrderEvent(ctx, order8)
 	mdb.Push(event8)
 
 	// Create new matching order
 	order9 := buildOrder("Normal3", types.SideSell, types.OrderTypeLimit, 100, 1, 0)
 	order9.Status = types.OrderStatusFilled
-	event9 := events.NewOrderEvent(ctx, order9)
+	event9 := newOrderEvent(ctx, order9)
 	mdb.Push(event9)
 
 	// Park pegged order
 	order10 := buildOrder("Pegged1", types.SideBuy, types.OrderTypeLimit, 99, 10, 10)
 	order10.PeggedOrder = &types.PeggedOrder{Reference: types.PeggedReferenceBestBid, Offset: num.NewUint(1)}
 	order10.Status = types.OrderStatusParked
-	event10 := events.NewOrderEvent(ctx, order10)
+	event10 := newOrderEvent(ctx, order10)
 	mdb.Push(event10)
 
 	md, err := mdb.GetMarketDepth(ctx, "M", 0)